调整创建默认配置代码
